\name{POEM_col}
\alias{POEM_col}
\title{Default POEM colour palette}
\description{
Load the default colour palette used in the POEM project for data visualisation.}
\usage{
POEM_col()}
\value{
A named vector with the default POEM colors.
}
\author{Benjamin M. Delory}
\examples{
colors<-POEM_col()
}
