\name{rhizotron_explorer}
\alias{rhizotron_explorer}
\title{Open (mini)rhizotron data explorer}
\description{
Opens a shiny app to manage (mini)rhizotron images and process segmented root images.}
\usage{
rhizotron_explorer()}
\value{
Opens a shiny app.
}
\author{Benjamin M. Delory}
\examples{
rhizotron_explorer()
}
